Image Resizing
Scaling operations increase or reduce the size of an image. 
The cv2.resize() function is used to resize an python image in OpenCV.
When the python image is resized, the interpolation method defines how the new pixels are computed. There are several interpolation techniques, each of which has its own quality vs. speed trade-offs.
It is important to note that resizing an image can reduce its quality. This is because the new pixels are calculated by interpolating between the existing pixels, and this can introduce some blurring.