Image Translation
Translating an image means shifting it within a given frame of reference that can be along the x-axis and y-axis.

To translate an image using OpenCV, we need to create a transformation matrix. This matrix is a 2×3 matrix that specifies the amount of translation in each direction.
The cv2.warpAffine() function is used to apply a transformation matrix to an image. It takes the following arguments:
The image to be transformed.
The transformation matrix.
The output image size.
The translation parameters are specified in the transformation matrix as the tx and ty elements. The tx element specifies the amount of translation in the x-axis, while the ty element specifies the amount of translation in the y-axis.